Experiment with MathJax and Highlight.js again - for #3
Got highlight.js to work when the FENCED_CODE extension in hoeown was enabled.
Currently, Math stuff is parsed badly by hoedown, but there may be stuff we can do.
Things look alright now as long as advanced stuff is wrapped in paragraph HTML tags.
But that's obviously not what you want in your markdown files..